server/requirements.txt
fastapi
uvicorn[standard]>=0.18.0
python-dotenv
bcrypt
pyjwt
openai>=1.0.0
sqlalchemy>=1.4,<2.0
psycopg2-binary==2.9.6
alembic>=1.7.0,<2.0
pydantic>=1.10,<2.0
bleach
requests
beautifulsoup4
pandas>=2.0.0,<3.0.0
openpyxl>=3.0.0,<4.0.0
